Moved AUTH_USER_MODEL tests to a separate django app.
