+++ tests/customuser/models.py
@@ -1,0 +1,1 @@
+from django.contrib.auth.tests.custom_user import CustomUser
