+++ tests/customuser/tests/__init__.py
@@ -1,0 +1,1 @@
+from customuser.tests import *
